Made EXR loader faster.  Updated FLTK2 to search for non debug libs in debug mode.  Removed warning from mrvSwizzleAudio.h.
Former-commit-id: 6e37d4e313b6cc88d3abd2a99f7e936daf084e43

+++ mrViewer/src/core/mrvSwizzleAudio.h
@@ -12,7 +12,7 @@ struct Swizzle
      
      inline Swizzle(void* data, unsigned data_size ) :
      ptr( (T*) data ),
-     last( data_size / 8*sizeof(T) )
+     last( data_size / unsigned(8*sizeof(T)) )
      {
      }
